Add long_name to the catalog
This entails opening netCDF files to get the long_name attribute; the current
implementation opens one file per variable name per component, but if a
variable is spread across multiple files it is assumed that the long_name does
not change.

Also created local copies of some of the other data in the catalog to make it
easier to reference between columns (e.g. storing path locally so I don't need
to access catalog['path'][-1] to get the most recent path value)